V3_Web/backend/requirements_v3.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
torch==2.3.1+cpu
torchvision==0.18.1+cpu
ultralytics==8.2.28
